Put the flushing backend in a separate file
...aids readability a bit.
